feat: update sse cmd
